RT-Ticket: 5716 RT-Update: correspond RT-Status: resolved
Move code which is not specific to parsing mime messages into
   RT::Email::Interface; this removes a large base of duplicated code,
   and some sketchy "methods" that were getting called with "" as self
   often

git-svn-id: svn+ssh://svn.bestpractical.com/svn/bps-public/rt/branches/3.2-RELEASE@1338 e417ac7c-1bcc-0310-8ffa-8f5827389a85

+++ lib/RT/I18N.pm
@@ -357,7 +357,6 @@ sub _FindOrGuessCharset {
 
 # }}}
 
-
 # {{{ _GuessCharset
 
 =head2 _GuessCharset STRING
@@ -423,6 +422,7 @@ sub SetMIMEHeadToEncoding {
     return if $charset eq $enc and $preserve_words;
 
     foreach my $tag ( $head->tags ) {
+        next unless $tag; # seen in wild: headers with no name
         my @values = $head->get_all($tag);
         $head->delete($tag);
         foreach my $value (@values) {
